Removing duplicity code, updating relate test
