+++ driver/src/data/record.ts
@@ -7,3 +7,13 @@ export const recordInternalProperties = [
   '@logicalName',
   '@key',
 ];
+
+export function exludeInternalPropertiesFromPayload(record: Record) {
+  const updatedRecord = { ...record } as Record;
+  Object.keys(record).forEach((key) => {
+    if (recordInternalProperties.includes(key)) {
+      delete updatedRecord[key];
+    }
+  });
+  return updatedRecord;
+}
