+++ driver/src/repositories/currentUserRecordRepository.ts
@@ -1,5 +1,5 @@
 import RecordRepository from './recordRepository';
-import Record, { recordInternalProperties } from '../data/record';
+import Record, { exludeInternalPropertiesFromPayload } from '../data/record';
 import { AssociateRequest } from '../requests';
 
 /**
@@ -52,7 +52,7 @@ export default class CurrentUserRecordRepository implements RecordRepository {
      */
   public async createRecord(logicalName: string, record: Record): Promise<Xrm.LookupValue> {
     return this.webApi.createRecord(logicalName,
-      CurrentUserRecordRepository.exludeInternalPropertiesFromPayload(record));
+      exludeInternalPropertiesFromPayload(record));
   }
 
   /**
@@ -65,7 +65,7 @@ export default class CurrentUserRecordRepository implements RecordRepository {
   public async upsertRecord(logicalName: string, record: Record): Promise<Xrm.LookupValue> {
     if (!record['@key']) {
       return this.webApi.createRecord(logicalName,
-        CurrentUserRecordRepository.exludeInternalPropertiesFromPayload(record));
+        exludeInternalPropertiesFromPayload(record));
     }
 
     const retrieveResponse = await this.webApi.retrieveMultipleRecords(
@@ -76,13 +76,13 @@ export default class CurrentUserRecordRepository implements RecordRepository {
     if (retrieveResponse.entities.length > 0) {
       const id = retrieveResponse.entities[0][`${logicalName}id`];
       await this.webApi.updateRecord(logicalName, id,
-        CurrentUserRecordRepository.exludeInternalPropertiesFromPayload(record));
+        exludeInternalPropertiesFromPayload(record));
 
       return { entityType: logicalName, id };
     }
 
     return this.webApi.createRecord(logicalName,
-      CurrentUserRecordRepository.exludeInternalPropertiesFromPayload(record));
+      exludeInternalPropertiesFromPayload(record));
   }
 
   /**
@@ -111,15 +111,5 @@ export default class CurrentUserRecordRepository implements RecordRepository {
     relationship: string,
   ): Promise<void> {
     this.webApi.execute(new AssociateRequest(primaryRecord, relatedRecords, relationship));
-  }
-
-  private static exludeInternalPropertiesFromPayload(record: Record) {
-    const updatedRecord = { ...record } as Record;
-    Object.keys(record).forEach((key) => {
-      if (recordInternalProperties.includes(key)) {
-        delete updatedRecord[key];
-      }
-    });
-    return updatedRecord;
   }
 }
